requirements.txt
pillow==12.1.1
numpy==2.4.2
scikit-image==0.26.0